+++ src/__init__.py
@@ -1,2 +1,3 @@
 from .util import *  # noqa
 from .discord_client import DiscordClient  # noqa
+from .trie import Trie  # noqa